# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "To-Do List API",
        "description": "Laravel asosida qurilgan oddiy To-Do List REST API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Barcha tasklarni qaytaradi",
                "operationId": "c641d93887fb50d557f18a7d11d0fd70",
                "responses": {
                    "200": {
                        "description": "Tasklar ro‘yxati"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Yangi task yaratadi",
                "operationId": "53feb2a5084021acb62923cf3cd7a873",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Kitob o‘qish"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "O‘zbek adabiyoti haqida"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task yaratildi"
                    },
                    "422": {
                        "description": "Validatsiya xatosi"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "ID bo‘yicha bitta taskni qaytaradi",
                "operationId": "23615b21f3984fd9361306821c838ad0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task ID raqami",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task topildi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Taskni yangilaydi",
                "operationId": "18e64b4b16cbee264fa84b775e6b7a91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Yangilanadigan task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Yangilangan sarlavha"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Yangilangan matn"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "done"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task yangilandi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Taskni o‘chiradi",
                "operationId": "2c379aab1f7cf16dbf084cee9978de75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "O‘chiriladigan task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task o‘chirildi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}